+++ dbt_city_cycles/models/marts/mart_hourly_patterns_summary.sql
@@ -5,7 +5,7 @@
 select
     location,
     hour_of_day,
-    count(*) as ride_count
-from {{ ref('unified_rides') }}
+    sum(ride_count) as ride_count
+from {{ ref('mart_hourly_rides') }}
 group by 1, 2
-order by 1, 2 
+order by 1, 2